feat: refactor directory schedule components and update documentation for metadata changes
