+++ apps/web/src/components/directories/detail/DirectoryTabs.tsx
@@ -66,7 +66,7 @@ export function DirectoryTabs({ directory }: DirectoryTabsProps) {
         {
             name: t('schedule'),
             href: `${ROUTES.DASHBOARD_DIRECTORY(directory.id)}/schedule`,
-            visible: Boolean(config?.metadata?.initial_prompt),
+            visible: Boolean(config),
             icon: (
                 <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                     <path
